Adding cakephp/elastic-search and mongodb/mongodb to the project folder
Basically, we don't have time to mess around with the PHP.ini, PEAR and other things on our own individual boxes, so we can't use composer.json to manage the packages right now.

Instead, we downloaded this stuff from the source and will eventually install this stuff with composer.json.